Allow bypassing FSA check for some actions
Some libraries that dispatch actions through our reducers do not
conform to the flux-standard-action spec, but `createReducer` will
throw a `NonStandardAction` when encountering those actions.

This PR adds allows `createReducer` to take an
`allowNonStandardActionIf` predicate that is applied to each action.
If the predicate returns true, we no longer check to see if it conforms
to the FSA spec.